+++ example/lib/main.dart
@@ -41,60 +41,92 @@ class _SearchFieldSampleState extends State<SearchFieldSample> {
   @override
   void initState() {
     cities = [
-      City('New York', 10001),
-      City('Los Angeles', 90001),
-      City('Chicago', 60601),
-      City('Houston', 77001),
-      City('Phoenix', 85001),
-      City('Philadelphia', 19101),
-      City('San Antonio', 78201),
-      City('San Diego', 92101),
-      City('Dallas', 75201),
-      City('San Jose', 95101),
+      City('New York', '10001'),
+      City('Los Angeles', '90001'),
+      City('Chicago', '60601'),
+      City('Houston', '77001'),
+      City('Phoenix', '85001'),
+      City('Philadelphia', '19101'),
+      City('San Antonio', '78201'),
+      City('San Diego', '92101'),
+      City('Dallas', '75201'),
+      City('San Jose', '95101'),
+      City('Austin', '73301'),
+      City('Jacksonville', '32099'),
+      City('Fort Worth', '76101'),
+      City('Columbus', '43201'),
+      City('Charlotte', '28201'),
+      City('San Francisco', '94101'),
+      City('Indianapolis', '46201'),
+      City('Seattle', '98101'),
+      City('Denver', '80201'),
+      City('Washington', '20001'),
+      City('Boston', '02101'),
     ].map(
       (City ct) {
         return SearchFieldListItem<City>(
           ct.name,
           value: ct.zip.toString(),
           item: ct,
-          child: Text(ct.name),
+          child: searchChild(ct, isSelected: false),
         );
       },
     ).toList();
-    selectedValue = cities[0];
     super.initState();
   }
 
+  Widget searchChild(City city, {bool isSelected = false}) => ListTile(
+        contentPadding: EdgeInsets.all(0),
+        title: Text(city.name,
+            style: TextStyle(color: isSelected ? Colors.green : null)),
+        trailing: Text('#${city.zip}'),
+      );
   var cities = <SearchFieldListItem<City>>[];
-  late SearchFieldListItem<City>? selectedValue;
+  SearchFieldListItem<City>? selectedValue;
   @override
   Widget build(BuildContext context) {
-    Widget searchChild(x, {bool isSelected = false}) => Padding(
-          padding: const EdgeInsets.symmetric(horizontal: 12),
-          child: Text(x,
-              style: TextStyle(
-                  fontSize: 18, color: isSelected ? Colors.green : null)),
-        );
     return Scaffold(
         appBar: AppBar(title: Text('Searchfield Demo')),
         body: Padding(
           padding: const EdgeInsets.all(8.0),
           child: Column(
-            mainAxisAlignment: MainAxisAlignment.center,
+            mainAxisAlignment: MainAxisAlignment.start,
             spacing: 20,
             children: [
+              TextField(
+                decoration: InputDecoration(
+                  labelText: 'Enter username',
+                  border: OutlineInputBorder(
+                    borderRadius: BorderRadius.all(Radius.circular(8)),
+                  ),
+                ),
+              ),
               SearchField(
+                suggestionsDecoration: SuggestionDecoration(
+                  borderRadius: BorderRadius.all(Radius.circular(2)),
+                  itemPadding: EdgeInsets.symmetric(horizontal: 16),
+                ),
                 hint: 'Search for a city or zip code',
-                // dynamicHeight: true,
                 maxSuggestionBoxHeight: 300,
                 onSuggestionTap: (SearchFieldListItem<City> item) {
                   setState(() {
                     selectedValue = item;
                   });
                 },
+                searchInputDecoration: SearchInputDecoration(
+                  prefixIcon: Icon(Icons.search),
+                  suffix: Icon(Icons.expand_more),
+                  border: OutlineInputBorder(
+                    borderRadius: BorderRadius.all(Radius.circular(8)),
+                  ),
+                ),
                 onSearchTextChanged: (searchText) {
                   if (searchText.isEmpty) {
-                    return cities;
+                    return cities
+                        .map((e) => e.copyWith(
+                            child: searchChild(e.item!,
+                                isSelected: e == selectedValue)))
+                        .toList();
                   }
                   // filter the list of cities by the search text
                   final filter = List<SearchFieldListItem<City>>.from(cities)
@@ -109,9 +141,8 @@ class _SearchFieldSampleState extends State<SearchFieldSample> {
                 selectedValue: selectedValue,
                 suggestions: cities
                     .map((e) => e.copyWith(
-                          child: searchChild(e.item!.name,
-                              isSelected: selectedValue == e),
-                        ))
+                        child: searchChild(e.item!,
+                            isSelected: e == selectedValue)))
                     .toList(),
                 suggestionState: Suggestion.expand,
               ),
@@ -130,6 +161,6 @@ class _SearchFieldSampleState extends State<SearchFieldSample> {
 
 class City {
   String name;
-  int zip;
+  String zip;
   City(this.name, this.zip);
 }